add script to test 'UTXOHNodeWallet' init() and update 'signTransaction'
