+++ .gitignore
@@ -5,4 +5,4 @@ cache
 artifacts
 
 #Env files
-.env
+.env*
